django_backend/requirements.txt
amqp==5.3.1
asgiref==3.9.0
beautifulsoup4==4.12.2
billiard==4.2.1
boto3==1.40.1
botocore==1.40.1
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.11
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.5.4
MarkupSafe==3.0.3
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
razorpay==2.0.0
redis==5.2.1
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
